Rename profile.rs -> profile_pane.rs
