+++ src/tui/view/component.rs
@@ -3,8 +3,8 @@
 mod help;
 mod misc;
 mod primary;
-mod profile;
 mod profile_list;
+mod profile_pane;
 mod recipe_list;
 mod recipe_pane;
 mod record_body;
