+++ src/tui/view/component/profile_pane.rs
